:sparkles: Add the ability to rename the release task.
The 'release' task may conflict with a user's chosen, existing 'release'
task. As such, users should be able to tell GulpSSHDeploy what they want
the release task to be called. Also, 'release' isn't as good of a name as
'deploy' in this context, so the default task name was changed, as well.